+++ src/views/Favorites.vue
@@ -13,7 +13,7 @@
                 </div>
                 <SkeletonCardList v-if="isLoading" :cols="colSizes" :limitRows="2" :dense="currentGridSize > 0" />
                 <VideoCardList
-                    :videos="sortedLive"
+                    :videos="live"
                     includeChannel
                     includeAvatar
                     :limitRows="2"
@@ -160,13 +160,6 @@ export default {
                 lg: 4 + this.currentGridSize,
                 xl: 5 + this.currentGridSize,
             };
-        },
-        sortedLive() {
-            return this.live.sort((a, b) => {
-                const dateA = new Date(a.available_at).getTime();
-                const dateB = new Date(b.available_at).getTime();
-                return dateA > dateB ? 1 : -1;
-            });
         },
     },
     methods: {