+++ src/views/Home.vue
@@ -27,7 +27,7 @@
                 <div class="text-h6">
                     {{ $t("views.home.recentVideosHeading") }}
                 </div>
-                <v-btn-toggle v-model="recentVideoFilter" mandatory dense>
+                <v-btn-toggle v-model="recentVideoFilter" mandatory dense color="secondary">
                     <v-btn value="all">
                         {{ $t("views.home.recentVideoToggles.all") }}
                     </v-btn>